fixing the type of '-' into '+'
The equation should be '+' not a '-'. I have double checked the math and I also notice that another student has made a similar observation. 

it's a minor typo, but it did confuse me enough to lose an hour just making me think if I have gotten bad at simple addition and subtraction.